Update SearchBooks.jsx
The saveBook mutation is called with the necessary variables corresponding to the book data. Context option is used to set the authorization header with the user's token. After a successful save, the book's ID is added to the savedBookIds state to update the UI accordingly. Error handling is performed in the catch block, logging out any errors that occur during the mutation.