Update SignupForm.jsx
useMutation hook is used to call the ADD_USER mutation when the form is submitted. The error object returned from useMutation is used to conditionally show an alert if the mutation fails. setShowAlert is called with true if there's an error in the mutation, which will display an error alert to the user. After a successful sign-up, the user's form data is reset to the initial state.